Reorganise user tests and test helpers
